project2/requirements.txt
sqlalchemy
python-dotenv
mysqlclient
mysql-connector-python